CREATED 2 VIEWPORTS AND IMPLEMENTED A HB BUTTON.
I decided to add a side-bar to the viewport to act as a menu for the content displayed in the main vieport (e.g. It could include actions like "Create new list" or "View deleted lists" etc) and implemented a functional soon-to-be "hamburger-icon" button. Which toggles the sibe-bar display on and off.

Main feautures:

- Created a menu-type column
- Added a button to toggle menu-like column
- Both viewports are screen size responsive

These feautires allow the user to be able to choose between both viewports, or simply just one (the main content).