FINALISED THE LAYOUT AND MOST OF THE STYLING.
The layout of the website has been completed. And the main features of the website have been added and can be seen.
These include:

- Search-bar to find lists by title.
- Side-bar menu to take actions such as creating new lists and being able to view the "starred" lists and the lists the user has deleted.
- Personalisation setting. So the user can customise the theme and change their name.

These changes have been added to make it easier for the user to naviagate the site with more ease and to slightly customise it, which will add a more personal feel.