"NEW LIST" FUNCTIONALITY IMPROVEMENT
Fixed issue with being able to create a new list when other lists were hidden (non-starred lists) which essentially created two different group of lists that were getting toggled on and off.
Now once the "Starred" link is clicked, the "New List" link is rendered unclickable and therefore denies the user form creating new lists while other lists are hidden.

This was made possible with the .on() and .off() methods.